enhancement(AppBar): renames
