+++ packages/studio-base/src/components/AppBar/AddPanel.stories.tsx
