+++ packages/studio-base/src/components/AppBar/CustomWindowControls.tsx
