+++ packages/studio-base/src/components/AppBar/EndTimestamp.stories.tsx
