+++ packages/studio-base/src/components/AppBar/EndTimestamp.tsx
