+++ packages/studio-base/src/components/AppBar/types.ts
